Created a new module pathops.py
